% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{adjustmentCalculation}
\alias{adjustmentCalculation}
\title{A Function to Convert the UI P-Value Adjustment into the Backend P-Value Adjustment}
\usage{
adjustmentCalculation(adjustment)
}
\arguments{
\item{adjustment}{A string character containing the adjustment to the P-value}
}
\description{
This function converts the P-value adjustment value from the UI into the value required by the backend
}
\examples{
adjustment <- adjustmentCalculation("Benjamini & Hochberg (False discovery rate)")
}
\author{
Guy Hunt
}
\keyword{GEO}
