% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{boxAndWhiskerPlot}
\alias{boxAndWhiskerPlot}
\title{A Function to Create a Box and Whisker Plot from an Expression Object}
\usage{
boxAndWhiskerPlot(geoAccessionCode = "", platform = "", ex)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}

\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot expression data into a Box and Whisker Plot
}
\examples{
fig <- boxAndWhiskerPlot(expressionData, "GSE18380", "GPL4694")
}
\author{
Guy Hunt
}
\keyword{GEO}
