% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateDifferentialGeneExpression}
\alias{calculateDifferentialGeneExpression}
\title{A Function to Calculate the Differential Gene EXpression between two groups}
\usage{
calculateDifferentialGeneExpression(
  gsms,
  limmaPrecisionWeights,
  forceNormalization,
  gset,
  ex
)
}
\arguments{
\item{gsms}{A string of intgers indicating which group a sample belongs to}

\item{limmaPrecisionWeights}{Whether to apply limma precision weights (vooma)}

\item{forceNormalization}{Whether to force normalization}

\item{gset}{The GEO object}

\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function calculates the differential expression for two groups
}
\examples{
fit2 <- calculateDifferentialGeneExpression(gsms, limmaPrecisionWeights, forceNormalization, gset, ex)
}
\author{
Guy Hunt
}
\keyword{GEO}
