% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateDT}
\alias{calculateDT}
\title{TBD}
\usage{
calculateDT(fit2, adjustment, significanceLevelCutOff)
}
\description{
TBD
}
\examples{
TBD
}
\author{
Guy Hunt
}
\keyword{GEO}
