% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateEachGroupsSamplesFromDataFrame}
\alias{calculateEachGroupsSamplesFromDataFrame}
\title{A Function to Calculate the Samples Selected in Each Group}
\usage{
calculateEachGroupsSamplesFromDataFrame(groupDataFrame)
}
\arguments{
\item{groupDataFrame}{A dataframe containing the row number and group selection}
}
\description{
This function calculates the GSMS object for differential expression from a dataframe of the groups
}
\examples{
TBD
}
\author{
Guy Hunt
}
\keyword{GEO}
