% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/backendDifferentialGeneExpressionDataFunctions.R
\name{calculateExclusiveColumns}
\alias{calculateExclusiveColumns}
\title{A Function to Create a List of Columns that Have Not Been Selected}
\usage{
calculateExclusiveColumns(columns, inputColumns)
}
\arguments{
\item{columns}{A list of all columns within the expression object}

\item{inputColumns}{A list of the columns that have been selected}
}
\description{
This function creates a list of columns that have not been selected in the UI
}
\examples{
column2 <- calculateExclusiveColumns(c("GSM455528", "GSM455541", "GSM455542", "GSM455543"), c("GSM455528", "GSM455541"))
}
\author{
Guy Hunt
}
\keyword{GEO}
