% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateFit2}
\alias{calculateFit2}
\title{A Function to Calculate the Differential Gene EXpression between two groups}
\usage{
calculateFit2(
  gsms,
  logTransformation,
  limmaPrecisionWeights,
  forceNormalization,
  knnTransformation,
  gset
)
}
\arguments{
\item{logTransformation}{Whether to auto-detect if log transformation is appropriate or to apply log transformation. Values can be "Auto-Detect" for auto detect, "Yes" to apply log transformation and "No" to not perform log transformation.}

\item{limmaPrecisionWeights}{Whether to apply limma precision weights (vooma)}

\item{forceNormalization}{Whether to force normalization}

\item{knnTransformation}{Whether to fill in missing values using Knn}
}
\description{
This function calculates the differential expression for two groups
}
\examples{
fit2 <- calculateFit2(gsms, logTransformation, limmaPrecisionWeights, forceNormalization, knnTransformation)
}
\author{
Guy Hunt
}
\keyword{GEO}
