% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{calculateKnnImpute}
\alias{calculateKnnImpute}
\title{A Function to Perform KNN Impute on an Expression Object}
\usage{
calculateKnnImpute(ex, knnTransformation)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{knnTransformation}{Whether to apply KNN impute. This can be "Yes" or "No"}
}
\description{
A function to impute missing expression data, using nearest neighbor averaging.
}
\examples{
knnDataInput <- calculateKnnImpute(dataInput, "Yes")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
