% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{calculatePrcompPca}
\alias{calculatePrcompPca}
\title{A Function to Perform Principle Component Analysis on an Expression Object}
\usage{
calculatePrcompPca(ex)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
A function to perform prcomp principle component analysis on an expression object.
}
\examples{
pcaDataInput <- calculatePrcompPca(knnDataInput)
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
