% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{calculatePrincompPca}
\alias{calculatePrincompPca}
\title{A Function to Perform Principle Component Analysis on an Expression Object}
\usage{
calculatePrincompPca(ex)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
A function to perform Princomp principle component analysis on an expression object.
}
\examples{
pcaPrincompDataInput <- calculatePrincompPca(knnDataInput)
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
