% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{correlationMatrixPlot}
\alias{correlationMatrixPlot}
\title{A Function to Create a Correlation Matrix that contains both the Correlations between Samples}
\usage{
correlationMatrixPlot(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot PCA expression results into an Scatter Plot contains both the Individuals and Variables
}
\examples{
fig <- correlationMatrixPlot(expressionData)
}
\author{
Guy Hunt
}
\keyword{GEO}
