% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{differentialGeneExpression}
\alias{differentialGeneExpression}
\title{TBD}
\usage{
differentialGeneExpression(
  gset,
  ex,
  gsms,
  limmaPrecisionWeights,
  forceNormalization
)
}
\description{
TBD
}
\examples{
TBD
}
\author{
Guy Hunt
}
\keyword{GEO}
