% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{expressionValueDistributionPlot}
\alias{expressionValueDistributionPlot}
\title{A Function to Create a Density Plot from an Expression Object}
\usage{
expressionValueDistributionPlot(geoAccessionCode = "", platform = "", ex)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}

\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot expression data into a Density Plot
}
\examples{
fig <- expressionValueDistributionPlot(expressionData, "GSE18380", "GPL4694")
}
\author{
Guy Hunt
}
\keyword{GEO}
