% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{extractColumns}
\alias{extractColumns}
\title{A Function to Extract Columns Names from an Expression Object}
\usage{
extractColumns(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function extracts the column names from an expression object
}
\examples{
columnNames <- extractColumns(expressionData)
}
\author{
Guy Hunt
}
\keyword{GEO}
