% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{extractExperimentInformation}
\alias{extractExperimentInformation}
\title{A GEO Function to Obtain the Experiment Information from a GEO Object}
\usage{
extractExperimentInformation(gset)
}
\arguments{
\item{gset}{The GEO object which can be obtained from the extractPlatformGset() function}
}
\description{
This function allows you to extract experiment information from a GEO object
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract the experiment information
experimentInformation <- extractExperimentInformation(gsetData)

}
\seealso{
\code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
