% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractGeneAnnotation}
\alias{extractGeneAnnotation}
\title{A GEO Function to Extract Gene Information from a GEO object}
\usage{
extractGeneAnnotation(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract the studies Gene information from a GEO object
}
\examples{
geneAnnotation <- extractGeneAnnotation(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
