% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{extractPlatformGset}
\alias{extractPlatformGset}
\title{A GEO Function to Obtain the GEO Object for a specific platform}
\usage{
extractPlatformGset(gset, platform)
}
\arguments{
\item{gset}{The GEO object which can be obtained from the getGeoObject() function}

\item{platform}{The platform code}
}
\description{
This function allows you to extract a platform's GEO object from a GEO object
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

}
\seealso{
\code{\link[=getGeoObject]{getGeoObject()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
