% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{extractPlatforms}
\alias{extractPlatforms}
\title{A GEO Function to Obtain the Available Platforms of a GEO Object}
\usage{
extractPlatforms(gset)
}
\arguments{
\item{gset}{The GEO object which can be obtained from the getGeoObject() function}
}
\description{
This function allows you to extract the platforms codes from a GEO object
}
\examples{
platforms <- extractPlatforms(allGset)
}
\seealso{
\code{\link[=getGeoObject]{getGeoObject()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
