% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractSampleDetails}
\alias{extractSampleDetails}
\title{A GEO Function to Extract Information on the Samples from a GEO object}
\usage{
extractSampleDetails(gset)
}
\arguments{
\item{gset}{The GEO object which can be obtained from the extractPlatformGset() function}
}
\description{
This function allows you to extract the studies sample information from a GEO object
}
\examples{
columnInfo <- extractSampleDetails(gsetData)
}
\seealso{
\code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
