% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractSampleInfo}
\alias{extractSampleInfo}
\title{A GEO Function to Extract Information on the Samples from a GEO object}
\usage{
extractSampleInfo(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract the studies sample information from a GEO object
}
\examples{
sampleInfo <- extractSampleInfo(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
