% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{extractSampleNames}
\alias{extractSampleNames}
\title{A Function to Extract the Sample/Columns Names from an Expression Object}
\usage{
extractSampleNames(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function extracts the sample/column names from an expression object
}
\examples{
columnNames <- extractSampleNames(expressionData)
}
\author{
Guy Hunt
}
\keyword{GEO}
