% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getColumnDetails}
\alias{getColumnDetails}
\title{A GEO Function to Extract Information on the Samples from a GEO object}
\usage{
getColumnDetails(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract the studies sample information from a GEO object
}
\examples{
columnInfo <- getColumnDetails(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
