% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getExperimentInformation}
\alias{getExperimentInformation}
\title{A GEO Function to Obtain the Experiment Information from a GEO Object}
\usage{
getExperimentInformation(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract experiment information from a GEO object
}
\examples{
experimentInformation <- getExperimentInformation(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
