% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getGeoData}
\alias{getGeoData}
\title{A GEO Data Sourcing Function}
\usage{
getGeoData(geoAccessionCode, platform)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{The study's platform}
}
\description{
This function allows you to source a GEO Object from GEO when you know the specific Geo Accession code and platform
}
\examples{
getGeoData("GSE18380", "GPL4694")
}
\author{
Guy Hunt
}
\keyword{GEO}
