% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getGset}
\alias{getGset}
\title{A GEO Data Sourcing Function}
\usage{
getGset(
  geoAccessionCode,
  GSEMatrix = TRUE,
  getGPL = TRUE,
  platformAnnotation = "NCBI generated"
)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{GSEMatrix}{A boolean telling GEOquery whether or not to use GSE Series Matrix files from GEO. The parsing of these files can be many orders-of-magnitude faster than parsing the GSE SOFT format files. Defaults to TRUE, meaning that the SOFT format parsing will not occur; set to FALSE if you for some reason need other columns from the GSE records.}

\item{getGPL}{A boolean defaulting to TRUE as to whether or not to download and include GPL information when getting a GSEMatrix file. You may want to set this to FALSE if you know that you are going to annotate your featureData using Bioconductor tools rather than relying on information provided through NCBI GEO. Download times can also be greatly reduced by specifying FALSE.}

\item{platformAnnotation}{A string defaulting to "NCBI generated" meaning true as to whether or not to use the Annotation GPL information. These files are nice to use because they contain up-to-date information remapped from Entrez Gene on a regular basis. However, they do not exist for all GPLs; in general, they are only available for GPLs referenced by a GDS. Input "Submitter supplied" for FALSE}
}
\description{
This function allows you to source a GEO Object from GEO when you only know the Geo Accession code
}
\examples{
allGset <- getGset("GSE18380", GSEMatrix=TRUE, getGPL=TRUE, platformAnnotation = "NCBI generated")
}
\author{
Guy Hunt
}
\keyword{GEO}
