% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getPlatformGset}
\alias{getPlatformGset}
\title{A GEO Function to Obtain the GEO Object for a specific platform}
\usage{
getPlatformGset(gset, platform)
}
\arguments{
\item{gset}{The GEO object}

\item{platform}{The platform code}
}
\description{
This function allows you to extract a platform's GEO object from a GEO object
}
\examples{
gsetData <- getPlatformGset(gset, platforms[1])
}
\author{
Guy Hunt
}
\keyword{GEO}
