% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{getPlatforms}
\alias{getPlatforms}
\title{A GEO Function to Obtain the Available Platforms of a GEO Object}
\usage{
getPlatforms(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract the platforms codes from a GEO object
}
\examples{
platforms <- getPlatforms(allGset)
}
\author{
Guy Hunt
}
\keyword{GEO}
