% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveDataVisualizationFunctions.R
\name{interactiveDensityPlot}
\alias{interactiveDensityPlot}
\title{A Function to Create an Interactive Density Plot from an Expression Object}
\usage{
interactiveDensityPlot(ex, geoAccessionCode, platform)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}
}
\description{
This function allows you to plot expression data into an interactive Density Plot
}
\examples{
fig <- interactiveDensityPlot(expressionData, "GSE18380", "GPL4694")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
