% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveDataVisualizationFunctions.R
\name{interactiveDesnityPlot}
\alias{interactiveDesnityPlot}
\title{A Function to Create an Interactive Density Plot from an Expression Object}
\usage{
interactiveDesnityPlot(ex, geoAccessionCode, platform)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}
}
\description{
This function allows you to plot expression data into an interactive Density Plot
}
\examples{
fig <- interactiveDesnityPlot(expressionData, "GSE18380", "GPL4694")
}
\author{
Guy Hunt
}
\keyword{GEO}
