% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveDifferentialGeneExpressionDataVisualizationFunctions.R
\name{interactiveHistogramPlot}
\alias{interactiveHistogramPlot}
\title{A Function to Create an Interactive Histogram of the P values from Differential Gene Expression Analysis}
\usage{
interactiveHistogramPlot(fit2, adjustment)
}
\arguments{
\item{fit2}{An object containing the results of differntial gene expression analysis which can be obtained from the calculateFit2() function}

\item{adjustment}{A character string containing the adjustment to P-values}
}
\description{
This function allows you to plot an interactive histogram of the P values from differential gene expression analysis
}
\examples{
fig <- interactiveHistogramPlot(fit2, adjustment)
}
\author{
Guy Hunt
}
\keyword{GEO}
