% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveDifferentialGeneExpressionDataVisualizationFunctions.R
\name{interactiveMeanDifferencePlot}
\alias{interactiveMeanDifferencePlot}
\title{A Function to Create an Interactive Mean Difference Plot of the log2 Fold Change Versus Average log2 Expression Values from Differential Gene Expression Analysis}
\usage{
interactiveMeanDifferencePlot(fit2, dT, ct)
}
\arguments{
\item{fit2}{An object containing the results of differential gene expression analysis which can be obtained from the calculateDifferentialGeneExpression() function}

\item{dT}{An object that summarises if each gene is unregulated, down regulated or has a similar level of expression which can be obtained from the calculateDifferentialGeneExpressionSummary() function}

\item{ct}{A integer indicating the column to select from the dT object}
}
\description{
This function allows you to plot an interactive mean difference plot of the log2 fold change versus average log2 expression values from differential gene expression analysis
}
\examples{
fig <- interactiveMeanDifferencePlot(fit2, dT, ct)
}
\author{
Guy Hunt
}
\keyword{GEO}
