% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactiveMeanVariancePlot}
\alias{interactiveMeanVariancePlot}
\title{A Function to Create an Interactive Mean Variance Plot from an Expression Object}
\usage{
interactiveMeanVariancePlot(ex, geoAccessionCode, gset)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{gset}{The GEO object which can be obtained from the extractPlatformGset() function}
}
\description{
This function allows you to plot expression data into an interactive Mean Variance Plot
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Remove all incomplete rows
naOmitInput <- calculateNaOmit(knnDataInput)

# Interactive Mean Variance Plot
fig <- interactiveMeanVariancePlot(naOmitInput, geoAccessionCode, gsetData)
fig

}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object, \code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
