% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveDataVisualizationFunctions.R
\name{interactivePcaScreePlot}
\alias{interactivePcaScreePlot}
\title{A Function to Create an Interactive Histogram of the Principle Components from the PCA outputs of an Expression Object}
\usage{
interactivePcaScreePlot(pcaData, geoAccessionCode)
}
\arguments{
\item{pcaData}{An object containing the results of PCA on a Geo Expression object which can be obtained from the pcaAnalysis() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot PCA expression results into an interactive Histogram of the Principle Components
}
\examples{
fig <- interactivePcaScreePlot(pcaData, "GSE18380")
}
\author{
Guy Hunt
}
\keyword{GEO}
