% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactivePrcompPcaIndividualsPlot}
\alias{interactivePrcompPcaIndividualsPlot}
\title{A Function to Create an Interactive Scatter Plot of the prcomp Principle Components Analysis of each of the Genes in an Expression Object}
\usage{
interactivePrcompPcaIndividualsPlot(pcaData, geoAccessionCode, gset)
}
\arguments{
\item{pcaData}{An object containing the results of PCA on a Geo Expression object which can be obtained from the calculatePrincompPca() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{gset}{The GEO object which can be obtained from the extractPlatformGset() function}
}
\description{
This function allows you to plot prcomp PCA expression results into an interactive Scatter Plot
}
\examples{
fig <- interactiveprcompPcaIndividualsPlot(pcaData, "GSE18380", gset)
}
\seealso{
\code{\link[=calculatePrcompPca]{calculatePrcompPca()}} for Princomp PCA expression object, \code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
