% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactivePrcompPcaScreePlot}
\alias{interactivePrcompPcaScreePlot}
\title{A Function to Create an Interactive Histogram of the Prcomp Principle Components from the PCA outputs of an Expression Object}
\usage{
interactivePrcompPcaScreePlot(pcaData, geoAccessionCode)
}
\arguments{
\item{pcaData}{An object containing the results of PCA on a Geo Expression object which can be obtained from the calculatePca() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot Prcomp PCA expression results into an interactive Histogram of the Principle Components
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Remove all incomplete rows
naOmitInput <- calculateNaOmit(knnDataInput)

# Perform Prcomp PCA analysis on KNN transformation expression data
pcaPrcompDataInput <- calculatePrcompPca(naOmitInput)

# Interactive Prcomp PCA Scree Plot
fig <- interactivePrcompPcaScreePlot(pcaPrcompDataInput, geoAccessionCode)
fig

}
\seealso{
\code{\link[=calculatePca]{calculatePca()}} for PCA expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
