% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactivePrcompPcaVariablesPlot}
\alias{interactivePrcompPcaVariablesPlot}
\title{A Function to Create an Interactive Scatter Plot of the Principle Components Analysis of each of the Samples in an Expression Object}
\usage{
interactivePrcompPcaVariablesPlot(pcaData, geoAccessionCode)
}
\arguments{
\item{pcaData}{An object containing the results of PCA on a Geo Expression object which can be obtained from the calculatePrincompPca() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot PCA expression results into an interactive Scatter Plot
}
\examples{
fig <- interactivePrcompPcaVariablesPlot(pcaData, "GSE18380")
}
\seealso{
\code{\link[=calculatePrcompPca]{calculatePrcompPca()}} for Princomp PCA expression object, \code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
