% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactivePrincompPcaVariablesPlot}
\alias{interactivePrincompPcaVariablesPlot}
\title{A Function to Create an Interactive Scatter Plot of the princomp Principle Components Analysis of each of the Samples in an Expression Object}
\usage{
interactivePrincompPcaVariablesPlot(pcaData, geoAccessionCode)
}
\arguments{
\item{pcaData}{An object containing the results of PCA on a Geo Expression object which can be obtained from the calculatePrincompPca() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot princomp PCA expression results into an interactive Scatter Plot
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Remove all incomplete rows
naOmitInput <- calculateNaOmit(knnDataInput)

# Perform Princomp PCA analysis on KNN transformation expression data#
pcaPrincompDataInput <- calculatePrincompPca(naOmitInput)

# Interactive Princomp PCA Variables Plot
fig <- interactivePrincompPcaVariablesPlot(pcaPrincompDataInput, geoAccessionCode)
fig

}
\seealso{
\code{\link[=calculatePrincompPca]{calculatePrincompPca()}} for Princomp PCA expression object, \code{\link[=extractPlatformGset]{extractPlatformGset()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
