% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveDifferentialGeneExpressionDataVisualizationFunctions.R
\name{interactiveQQPlot}
\alias{interactiveQQPlot}
\title{A Function to Create an Interactive QQ Plot of the Quantiles of a Data Sample Against the Theoretical Quantiles of a Student's T Distribution from Differential Gene Expression Analysis}
\usage{
interactiveQQPlot(fit2, dT, ct)
}
\arguments{
\item{fit2}{An object containing the results of differential gene expression analysis which can be obtained from the calculateDifferentialGeneExpression() function}

\item{dT}{An object that summarises if each gene is unregulated, down regulated or has a similar level of expression which can be obtained from the calculateDifferentialGeneExpressionSummary() function}

\item{ct}{A integer indicating the column to select from the dT object}
}
\description{
This function allows you to plot an interactive QQ plot of the quantiles of a data sample against the theoretical quantiles of a Student's t distribution from differential gene expression analysis
}
\examples{
fig <- interactiveQQPlot(fit2, dT, ct)
}
\author{
Guy Hunt
}
\keyword{GEO}
