% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/interactiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{interactiveThreeDDesnityPlot}
\alias{interactiveThreeDDesnityPlot}
\title{A Function to Create an Interactive Three Dimensional Density Plot from an Expression Object}
\usage{
interactiveThreeDDesnityPlot(ex, geoAccessionCode, platform)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}
}
\description{
This function allows you to plot expression data into an interactive Three Dimensional Density Plot
}
\examples{
fig <- interactiveThreeDDesnityPlot(expressionData, "GSE18380", "GPL4694")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
