% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{isLogTransformAutoApplied}
\alias{isLogTransformAutoApplied}
\title{A Function to Determine if Log Transformation should Automatically be Applied}
\usage{
isLogTransformAutoApplied(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to determine if log transformation should be performed on an expression objects
}
\examples{
autoLogInformation <- isLogTransformAutoApplied(expressionData)
}
\author{
Guy Hunt
}
\keyword{GEO}
