% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{knnDataTransformation}
\alias{knnDataTransformation}
\title{A Function to Perform KNN Impute on an Expression Object}
\usage{
knnDataTransformation(ex, knnTransformation)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}

\item{knnTransformation}{Whether to apply KNN impute. This can be "Yes" or "No"}
}
\description{
A function to impute missing expression data, using nearest neighbor averaging.
}
\examples{
knnDataInput <- knnDataTransformation(dataInput, knnTransformation)
}
\author{
Guy Hunt
}
\keyword{GEO}
