% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{loadApp}
\alias{loadApp}
\title{A Function to Load the GEOexplorer Shiny App}
\usage{
loadApp()
}
\description{
This function loads the GEOexplorer Shiny App.
The GEOexplorer Shiny App extends GEO2R's functionalities by enabling a richer set of analysis and graphics to be performed/generated from the gene expression data.
}
\examples{
app <- loadApp()
}
\author{
Guy Hunt
}
\keyword{GEO}
