% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{naOmitTransformation}
\alias{naOmitTransformation}
\title{A Function to Removes Rows in an Expression Object that Contain Null Values}
\usage{
naOmitTransformation(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
A function to perform remove rows that contain a null value from an expression object.
}
\examples{
naOmitInput <- naOmitTransformation(knnDataInput)
}
\author{
Guy Hunt
}
\keyword{GEO}
