% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{nonInteractiveBoxAndWhiskerPlot}
\alias{nonInteractiveBoxAndWhiskerPlot}
\title{A Function to Create a Box and Whisker Plot from an Expression Object}
\usage{
nonInteractiveBoxAndWhiskerPlot(ex, geoAccessionCode = "", platform = "")
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot expression data into a Box and Whisker Plot
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Non-Interactive Box-and-Whisker Plot
fig <- nonInteractiveBoxAndWhiskerPlot(ex = knnDataInput, geoAccessionCode = geoAccessionCode, platform = platform)

}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
