% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{nonInteractiveDensityPlot}
\alias{nonInteractiveDensityPlot}
\title{A Function to Create a Density Plot from an Expression Object}
\usage{
nonInteractiveDensityPlot(ex, geoAccessionCode = "", platform = "")
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}
}
\description{
This function allows you to plot expression data into a Density Plot
}
\examples{
fig <- nonInteractiveDensityPlot(expressionData, "GSE18380", "GPL4694")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
