% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveDifferentialGeneExpressionDataVisualizationFunctions.R
\name{nonInteractiveHistogramPlot}
\alias{nonInteractiveHistogramPlot}
\title{A Function to Create a Histogram of the P values from Differential Gene Expression Analysis}
\usage{
nonInteractiveHistogramPlot(fit2, adjustment)
}
\arguments{
\item{fit2}{An object containing the results of differntial gene expression analysis which can be obtained from the calculateDifferentialGeneExpression() function}

\item{adjustment}{A character string containing the adjustment to P-values}
}
\description{
This function allows you to plot a histogram of the P values from differential gene expression analysis
}
\examples{
fig <- nonInteractiveHistogramPlot(fit2, adjustment)
}
\seealso{
\code{\link[=calculateDifferentialGeneExpression]{calculateDifferentialGeneExpression()}} for differential gene expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
