% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{noninteractiveMeanDifferencePlot}
\alias{noninteractiveMeanDifferencePlot}
\title{A Function to Create a Mean Difference Plot of the log2 Fold Change Versus Average log2 Expression Values from Differential Gene Expression Analysis}
\usage{
noninteractiveMeanDifferencePlot(fit2, dT, ct)
}
\arguments{
\item{fit2}{An object containing the results of differential gene expression analysis which can be obtained from the calculateFit2() function}

\item{dT}{An object that summarises if each gene is unregulated, down regulated or has a similar level of expression which can be obtained from the calculateDT() function}

\item{ct}{A integer indicating the column to select from the dT object}
}
\description{
This function allows you to plot a mean difference plot of the log2 fold change versus average log2 expression values from differential gene expression analysis
}
\examples{
fig <- noninteractiveMeanDifferencePlot(fit2, dT, ct)
}
\author{
Guy Hunt
}
\keyword{GEO}
