% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{nonInteractiveMeanVariancePlot}
\alias{nonInteractiveMeanVariancePlot}
\title{A Function to Create a Mean Variance Plot from an Expression Object}
\usage{
nonInteractiveMeanVariancePlot(ex, geoAccessionCode = "", platform = "")
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}
}
\description{
This function allows you to plot expression data into a Mean Variance Plot
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Remove all incomplete rows
naOmitInput <- calculateNaOmit(knnDataInput)

# Non-Interactive Mean Variance Plot
fig <- nonInteractiveMeanVariancePlot(naOmitInput, geoAccessionCode)

}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
