% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{nonInteractivePcaIndividualsPlot}
\alias{nonInteractivePcaIndividualsPlot}
\title{A Function to Create an Individuals Scatter Plot from the PCA outputs of an Expression Object}
\usage{
nonInteractivePcaIndividualsPlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the calculatePrincompPca() function}
}
\description{
This function allows you to plot PCA expression results into an Individuals Scatter Plot
}
\examples{
fig <- nonInteractivePcaIndividualsPlot(pcaPrincompDataInput)
}
\seealso{
\code{\link[=calculatePrincompPca]{calculatePrincompPca()}} for Princomp Pca expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
