% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{nonInteractivePcaScreePlot}
\alias{nonInteractivePcaScreePlot}
\title{A Function to Create a Histogram of the Principle Components from the PCA outputs of an Expression Object}
\usage{
nonInteractivePcaScreePlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the calculatePrincompPca() function}
}
\description{
This function allows you to plot PCA expression results into a Histogram of the Principle Components
}
\examples{
fig <- nonInteractivePcaScreePlot(pcaPrincompDataInput)
}
\seealso{
\code{\link[=calculatePrincompPca]{calculatePrincompPca()}} for Princomp Pca expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
