% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{nonInteractivePcaVariablesPlot}
\alias{nonInteractivePcaVariablesPlot}
\title{A Function to Create an Variables Scatter Plot from the PCA outputs of an Expression Object}
\usage{
nonInteractivePcaVariablesPlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the calculatePrincompPca() function}
}
\description{
This function allows you to plot PCA expression results into an Variables Scatter Plot
}
\examples{
fig <- nonInteractivePcaVariablesPlot(pcaPrincompDataInput)
}
\author{
Guy Hunt
}
\keyword{GEO}
