% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveDifferentialGeneExpressionDataVisualizationFunctions.R
\name{nonInteractiveQQPlot}
\alias{nonInteractiveQQPlot}
\title{A Function to Create a QQ Plot of the Quantiles of a Data Sample Against the Theoretical Quantiles of a Student's T Distribution from Differential Gene Expression Analysis}
\usage{
nonInteractiveQQPlot(fit2)
}
\arguments{
\item{fit2}{An object containing the results of differential gene expression analysis which can be obtained from the calculateDifferentialGeneExpression() function}
}
\description{
This function allows you to plot a QQ plot of the quantiles of a data sample against the theoretical quantiles of a Student's t distribution from differential gene expression analysis
}
\examples{
fig <- nonInteractiveQQPlot(fit2)
}
\seealso{
\code{\link[=calculateDifferentialGeneExpression]{calculateDifferentialGeneExpression()}} for differential gene expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
