% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{nonInteractiveVennDiagramPlot}
\alias{nonInteractiveVennDiagramPlot}
\title{A Function to Plot a Venn Diagram with the Number of Genes that were and were not Differentially Expressed}
\usage{
nonInteractiveVennDiagramPlot(dT)
}
\arguments{
\item{dT}{An object that summarises if each gene is unregulated, down regulated or has a similar level of expression which can be obtained from the calculateDifferentialGeneExpressionSummary() function}
}
\description{
This function creates a venndigram containing the number of genes that were and were not differentially expressed
}
\examples{
fig <- nonInteractiveVennDiagramPlot(dT)
}
\seealso{
\code{\link[=calculateDifferentialGeneExpressionSummary]{calculateDifferentialGeneExpressionSummary()}} for differential gene expression summary object
}
\author{
Guy Hunt
}
\keyword{GEO}
