% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransformationFunctions.R
\name{pcaAnalysis}
\alias{pcaAnalysis}
\title{A Function to Perform Principle Component Analysis on an Expression Object}
\usage{
pcaAnalysis(ex)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
A function to perform prcomp principle component analysis on an expression object.
}
\examples{
pcaDataInput <- pcaAnalysis(knnDataInput)
}
\author{
Guy Hunt
}
\keyword{GEO}
