% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{pcaIndividualsPlot}
\alias{pcaIndividualsPlot}
\title{A Function to Create an Individuals Scatter Plot from the PCA outputs of an Expression Object}
\usage{
pcaIndividualsPlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the pcaPrincompAnalysis() function}
}
\description{
This function allows you to plot PCA expression results into an Individuals Scatter Plot
}
\examples{
fig <- pcaIndividualsPlot(pcaPrincompDataInput)
}
\author{
Guy Hunt
}
\keyword{GEO}
