% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{pcaScreePlot}
\alias{pcaScreePlot}
\title{A Function to Create a Histogram of the Principle Components from the PCA outputs of an Expression Object}
\usage{
pcaScreePlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the pcaPrincompAnalysis() function}
}
\description{
This function allows you to plot PCA expression results into a Histogram of the Principle Components
}
\examples{
fig <- pcaScreePlot(pcaPrincompDataInput)
}
\author{
Guy Hunt
}
\keyword{GEO}
