% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUiComponents.R
\name{sourceDatasetInformationUi}
\alias{sourceDatasetInformationUi}
\title{A Function to Return the Dataset Information Ui Component}
\usage{
sourceDatasetInformationUi()
}
\description{
A Function to Return the Dataset Information Ui Component
}
\author{
Guy Hunt
}
