% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUiComponents.R
\name{sourceDifferentialGeneExpressionAnalysisUi}
\alias{sourceDifferentialGeneExpressionAnalysisUi}
\title{A Function to Return the Differential Gene Expression Analysis Ui Component}
\usage{
sourceDifferentialGeneExpressionAnalysisUi()
}
\description{
A Function to Return the Differential Gene Expression Analysis Ui Component
}
\examples{
sourceDifferentialGeneExpressionAnalysisUi()
}
\author{
Guy Hunt
}
