% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUiComponents.R
\name{sourceExploratoryDataAnalysisUi}
\alias{sourceExploratoryDataAnalysisUi}
\title{A Function to Return the Exploratory Data Analysis Ui Component}
\usage{
sourceExploratoryDataAnalysisUi()
}
\description{
A Function to Return the Exploratory Data Analysis Ui Component
}
\author{
Guy Hunt
}
