% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServerComponents.R
\name{sourceServer}
\alias{sourceServer}
\title{A Function to Return the Server Component}
\usage{
sourceServer(input, output, session)
}
\description{
A Function to Return the Server Component
}
\examples{
sourceServer()
}
\author{
Guy Hunt
}
