% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUiComponents.R
\name{sourceSideBarUi}
\alias{sourceSideBarUi}
\title{A Function to Return the Side Bar Ui Component}
\usage{
sourceSideBarUi()
}
\description{
A Function to Return the Side Bar Ui Component
}
\examples{
sourceSideBarUi()
}
\author{
Guy Hunt
}
