% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{topDifferentiallyExpressedGenesTable}
\alias{topDifferentiallyExpressedGenesTable}
\title{TBD}
\usage{
topDifferentiallyExpressedGenesTable(fit2, adjustment)
}
\description{
TBD
}
\examples{
TBD
}
\author{
Guy Hunt
}
\keyword{GEO}
