% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{umapPlot}
\alias{umapPlot}
\title{A Function to Create a UMAP Plot from Expression Object}
\usage{
umapPlot(geoAccessionCode = "", platform = "", ex, knn)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{The study's platform}

\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot expression data into a UMAP Plot
}
\examples{
fig <- umapPlot(expressionData, "GSE18380", "GPL4694")
}
\author{
Guy Hunt
}
\keyword{GEO}
