% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{adjustmentCalculation}
\alias{adjustmentCalculation}
\title{TBD}
\usage{
adjustmentCalculation(adjustment)
}
\description{
TBD
}
\examples{
TBD
}
\author{
Guy Hunt
}
\keyword{GEO}
