% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{calculateAutoLogTransformApplication}
\alias{calculateAutoLogTransformApplication}
\title{A Function to Determine if Log Transformation should Automatically be Applied}
\usage{
calculateAutoLogTransformApplication(ex)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to determine if log transformation should be performed on an expression objects
}
\examples{
autoLogInformation <- calculateAutoLogTransformApplication(expressionData)
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
