% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateDifferentialGeneExpressionSummary}
\alias{calculateDifferentialGeneExpressionSummary}
\title{A Function to Create an Object Containing if Each Gene is Unregulated, Down Regulated or has a Similar Level of Expression between the Groups}
\usage{
calculateDifferentialGeneExpressionSummary(
  fit2,
  adjustment,
  significanceLevelCutOff
)
}
\arguments{
\item{fit2}{An object containing the differentially expressed genes analysis that can be obtained from the calculateDifferentialGeneExpression() function}

\item{adjustment}{A string character containing the adjustment to the P-value. The values can be: "fdr", "BY", "bonferroni", "hochberg", "holm", "hommel" or "none"}

\item{significanceLevelCutOff}{A float indicating the P-value cutoff. The values can be between 0 and 1}
}
\description{
This function creates an object containing if each gene is unreguate, downregulated or not
}
\examples{
dT <- calculateDifferentialGeneExpressionSummary(fit2, "fdr", 0.05)
}
\seealso{
\code{\link[=calculateDifferentialGeneExpression]{calculateDifferentialGeneExpression()}} for differential gene expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
