% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/backendDifferentialGeneExpressionDataFunctions.R
\name{calculateEachGroupsSamples}
\alias{calculateEachGroupsSamples}
\title{A Function to Calculate the Samples Selected in Each Group}
\usage{
calculateEachGroupsSamples(columnNames, group1, group2)
}
\arguments{
\item{columnNames}{All the sample names in the expression object which can be obtained from the extractSampleNames() function}

\item{group1}{The sample names in group 1. This must not contain any sample names that are in group2}

\item{group2}{The sample names in group 2. This must not contain any sample names that are in group1}
}
\description{
This function calculates the GSMS object for differential expression from the sample names and samples in each group
}
\examples{
gsms <- calculateEachGroupsSamples(columnNames,c("GSM455528", "GSM455541", "GSM455542", "GSM455543", "GSM455578", "GSM455610", "GSM455782"), c("GSM455783", "GSM455784", "GSM455785", "GSM455786", "GSM455787"))
}
\seealso{
\code{\link[=extractSampleNames]{extractSampleNames()}} for all the sample names
}
\author{
Guy Hunt
}
\keyword{GEO}
