% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/backendDifferentialGeneExpressionDataFunctions.R
\name{calculateEachGroupsSamples}
\alias{calculateEachGroupsSamples}
\title{A Function to Calculate the Samples Selected in Each Group}
\usage{
calculateEachGroupsSamples(columnNames, group1, group2)
}
\arguments{
\item{columnNames}{All the sample names in the expression object which can be obtained from the extractSampleNames() function}

\item{group1}{The sample names in group 1. This must not contain any sample names that are in group2}

\item{group2}{The sample names in group 2. This must not contain any sample names that are in group1}
}
\description{
This function calculates the GSMS object for differential expression from the sample names and samples in each group
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Extract experimental condition/sample names
columnNames <- extractSampleNames(expressionData)

# Define Groups
numberOfColumns <- length(columnNames)
numberOfColumns <- numberOfColumns + 1
halfNumberOfColumns <- ceiling(numberOfColumns/2)
i <- 0
group1 <- c()
group2 <- c()
for (name in columnNames) {
if (i < halfNumberOfColumns) {
    group1 <- c(group1, name)
        i <- i +1
          } else {
          group2 <- c(group2, name)
          i <- i +1
          }
          }

# Select columns in group2
column2 <- calculateExclusiveColumns(columnNames, group1)

# Calculate gsms
gsms <- calculateEachGroupsSamples(columnNames,group1, group2)

}
\seealso{
\code{\link[=extractSampleNames]{extractSampleNames()}} for all the sample names
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
