% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{calculateKnnImpute}
\alias{calculateKnnImpute}
\title{A Function to Perform KNN Impute on an Expression Object}
\usage{
calculateKnnImpute(ex, knnTransformation)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{knnTransformation}{Whether to apply KNN impute. This can be "Yes" or "No"}
}
\description{
A function to impute missing expression data, using nearest neighbor averaging.
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
