% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialGeneExpressionAnalysis.R
\name{calculateTopDifferentiallyExpressedGenes}
\alias{calculateTopDifferentiallyExpressedGenes}
\title{A Function to Create a Table of the Top Differentially Expressed Genes}
\usage{
calculateTopDifferentiallyExpressedGenes(fit2, adjustment)
}
\arguments{
\item{fit2}{An object containing the differentially expressed genes analysis that can be obtained from the calculateDifferentialGeneExpression() function}

\item{adjustment}{A string character containing the adjustment to the P-value. The values can be: "fdr", "BY", "bonferroni", "hochberg", "holm", "hommel" or "none"}
}
\description{
This function creates a table of the top differentially expressed genes
}
\examples{
tT <- calculateTopDifferentiallyExpressedGenes(fit2, "fdr")
}
\seealso{
\code{\link[=calculateDifferentialGeneExpression]{calculateDifferentialGeneExpression()}} for differential gene expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
