% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{convertExperimentInformation}
\alias{convertExperimentInformation}
\title{A GEO Function to Convert the Experiment Information Object into HTML}
\usage{
convertExperimentInformation(experimentData)
}
\arguments{
\item{experimentData}{The experiment object obtained from the extractExperimentInformation() function}
}
\description{
This function allows you to convert experiment information into HTML
}
\examples{
convertExperimentInformation <- extractExperimentInformation(experimentInformation)
}
\author{
Guy Hunt
}
\keyword{GEO}
