% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backendExploratoryDataAnalysisDataFunctions.R
\name{convertExperimentInformation}
\alias{convertExperimentInformation}
\title{A GEO Function to Convert the Experiment Information Object into HTML}
\usage{
convertExperimentInformation(experimentData)
}
\arguments{
\item{experimentData}{The experiment object obtained from the extractExperimentInformation() function}
}
\description{
This function allows you to convert experiment information into HTML
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract the experiment information
experimentInformation <- extractExperimentInformation(gsetData)

# Extract Experiment Information
extractedExperimentInformation <- convertExperimentInformation(experimentInformation)

}
\seealso{
\code{\link[=extractExperimentInformation]{extractExperimentInformation()}} for GEO object
}
\author{
Guy Hunt
}
\keyword{GEO}
