% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractExperimentInformation}
\alias{extractExperimentInformation}
\title{A GEO Function to Obtain the Experiment Information from a GEO Object}
\usage{
extractExperimentInformation(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract experiment information from a GEO object
}
\examples{
experimentInformation <- extractExperimentInformation(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
