% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractExperimentInformation}
\alias{extractExperimentInformation}
\title{A GEO Function to Convert the Experiment Information Object into HTML}
\usage{
extractExperimentInformation(experimentData)
}
\arguments{
\item{experimentData}{The experiment object obtained from the getExperimentInformation() function}
}
\description{
This function allows you to convert experiment information into HTML
}
\examples{
extractedExperimentInformation <- extractExperimentInformation(experimentInformation)
}
\author{
Guy Hunt
}
\keyword{GEO}
