% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoIntegrationFunctions.R
\name{extractExpressionData}
\alias{extractExpressionData}
\title{A GEO Function to Extract Expression Data from a GEO object}
\usage{
extractExpressionData(gset)
}
\arguments{
\item{gset}{The GEO object}
}
\description{
This function allows you to extract the studies expression object from a GEO object
}
\examples{
expressionData <- extractExpressionData(gsetData)
}
\author{
Guy Hunt
}
\keyword{GEO}
