% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/backendDifferentialGeneExpressionDataFunctions.R
\name{extractSampleNames}
\alias{extractSampleNames}
\title{A Function to Extract the Sample/Columns Names from an Expression Object}
\usage{
extractSampleNames(ex)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function extracts the sample/column names from an expression object
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Extract experimental condition/sample names
columnNames <- extractSampleNames(expressionData)

}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
\keyword{internal}
