% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveDataVisualizationFunctions.R
\name{interactiveMeanVariancePlot}
\alias{interactiveMeanVariancePlot}
\title{A Function to Create an Interactive Mean Variance Plot from an Expression Object}
\usage{
interactiveMeanVariancePlot(ex, geoAccessionCode, gset)
}
\arguments{
\item{ex}{A GEO expression object which can be obtained from the extractExpressionData() function}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{gset}{The GEO object}
}
\description{
This function allows you to plot expression data into an interactive Mean Variance Plot
}
\examples{
fig <- interactiveMeanVariancePlot(expressionData, "GSE18380", gset)
}
\author{
Guy Hunt
}
\keyword{GEO}
