% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveDataVisualizationFunctions.R
\name{interactiveUmapPlot}
\alias{interactiveUmapPlot}
\title{A Function to Create an Interactive UMAP Plot from an Expression Object}
\usage{
interactiveUmapPlot(ex, knn, geoAccessionCode)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}

\item{knn}{A integer representing the number of k nearest neighbor's to use}

\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}
}
\description{
This function allows you to plot expression data into an interactive UMAP Plot
}
\examples{
fig <- interactiveUmapPlot(expressionData, 2, "GSE18380")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
