% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{loadApp}
\alias{loadApp}
\title{A Function to Load the Shiny App}
\usage{
loadApp()
}
\description{
This function loads the Shiny App
}
\examples{
fit2 <- calculateFit2(gsms, logTransformation, limmaPrecisionWeights, forceNormalization, knnTransformation)
}
\author{
Guy Hunt
}
\keyword{GEO}
