% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{nonInteractiveCorrelationMatrixPlot}
\alias{nonInteractiveCorrelationMatrixPlot}
\title{A Function to Create a Correlation Matrix that contains both the Correlations between Samples}
\usage{
nonInteractiveCorrelationMatrixPlot(ex)
}
\arguments{
\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot a heatmap of the correlations between experimental conditions
}
\examples{
fig <- nonInteractiveCorrelationMatrixPlot(expressionData)
}
\author{
Guy Hunt
}
\keyword{GEO}
