% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/nonInteractiveExploratoryDataAnalysisDataVisualizationFunctions.R
\name{nonInteractivePcaVariablesPlot}
\alias{nonInteractivePcaVariablesPlot}
\title{A Function to Create an Variables Scatter Plot from the PCA outputs of an Expression Object}
\usage{
nonInteractivePcaVariablesPlot(pcaEx)
}
\arguments{
\item{pcaEx}{A PCA object which can be obtained from the calculatePrincompPca() function}
}
\description{
This function allows you to plot PCA expression results into an Variables Scatter Plot
}
\examples{
# Get the GEO data for all platforms
geoAccessionCode <- "GSE18388"
allGset <- getGeoObject(geoAccessionCode)

# Extract platforms
platforms <- extractPlatforms(allGset)
platform <- platforms[1]

# Extract the GEO2R data from the specified platform
gsetData <- extractPlatformGset(allGset, platform)

# Extract expression data
expressionData <- extractExpressionData(gsetData)

# Apply log transformation to expression data if necessary
logTransformation <- "Auto-Detect"
dataInput <- calculateLogTransformation(expressionData, logTransformation)

# Perform KNN transformation on log expression data if necessary
knnDataInput <- calculateKnnImpute(dataInput, "Yes")

# Remove all incomplete rows
naOmitInput <- calculateNaOmit(knnDataInput)

# Perform Princomp PCA analysis on KNN transformation expression data
pcaPrincompDataInput <- calculatePrincompPca(naOmitInput)

# Non-Interactive Princomp PCA Variables Plot
fig <- nonInteractivePcaVariablesPlot(pcaPrincompDataInput)
fig

}
\seealso{
\code{\link[=calculatePrincompPca]{calculatePrincompPca()}} for Princomp Pca expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
