% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisualizationFunctions.R
\name{nonInteractiveUmapPlot}
\alias{nonInteractiveUmapPlot}
\title{A Function to Create a UMAP Plot from an Expression Object}
\usage{
nonInteractiveUmapPlot(geoAccessionCode = "", platform = "", ex, knn)
}
\arguments{
\item{geoAccessionCode}{A character string representing a GEO object for download and parsing}

\item{platform}{A character string representing the study's platform}

\item{ex}{The GEO expression object which can be obtained from the extractExpressionData() function}
}
\description{
This function allows you to plot expression data into a UMAP Plot
}
\examples{
fig <- nonInteractiveUmapPlot(expressionData, "GSE18380", "GPL4694")
}
\seealso{
\code{\link[=extractExpressionData]{extractExpressionData()}} for expression object
}
\author{
Guy Hunt
}
\keyword{GEO}
